U-NET_pytorch/requirements.txt
hydra-core
Pillow
albumentations
torchvision==0.9.1+cu111
torch==1.8.1+cu111
colorama
